Cleaner logging, remove global function.
