+++ ExampleVisionOS/MeshViewModel.swift
@@ -20,6 +20,26 @@ class MeshViewModel: ObservableObject {
     /// The RealityView instance has added the content.
     public var contentAdded = false
 
+    /// A demonstration Mesh generated by Euclid
+     private let euclidMesh: Mesh = {
+        let start = CFAbsoluteTimeGetCurrent()
+
+        // create some geometry using Euclid
+        let cube = Mesh.cube(size: 0.8, material: Color.red)
+        let sphere = Mesh.sphere(slices: 120, material: CGImage.checkerboard())
+        let mesh = cube.subtracting(sphere).makeWatertight()
+
+        print("Time:", CFAbsoluteTimeGetCurrent() - start)
+        print("Polygons:", mesh.polygons.count)
+        print("Triangles:", mesh.triangulate().polygons.count)
+        print("Watertight:", mesh.isWatertight)
+
+        // Sleep here to simulate a long-running mesh construction.
+        sleep(5)
+
+        return mesh
+    }()
+
     func prepareContent() {
         Task.init {
             do {
@@ -38,7 +58,7 @@ class MeshViewModel: ObservableObject {
                 entity = demoBoxEntity
                 contentReady = true
             } catch {
-                print("nope")
+                print(#function, "failed to update content")
             }
         }
     }
